+++ src/main/java/com/example/fanmon_be/domain/chat/dao/ChatDAO.java
@@ -6,7 +6,6 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
-import java.util.List;
 import java.util.UUID;
 @Repository
 public interface ChatDAO extends JpaRepository<Chat, UUID> {
