+++ src/main/java/com/example/fanmon_be/domain/chat/service/ChatService.java
@@ -6,7 +6,6 @@ import com.example.fanmon_be.domain.user.entity.User;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
 import java.util.UUID;
 
 @Service
