+++ src/main/java/com/example/fanmon_be/domain/chat/service/SubscribeService.java
@@ -5,7 +5,6 @@ import com.example.fanmon_be.domain.chat.entity.Subscribe;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
 import java.util.UUID;
 
 @Service
@@ -26,9 +25,5 @@ public class SubscribeService {
             return subscribe;
         }
         return null;
-    }
-
-    public List<Subscribe> mySubscriptionList(UUID useruuid){
-        return dao.findByUserUseruuid(useruuid);
     }
 }
